refractor tests with SimpleTestPost
