+++ app/src/test/java/com/github/versus/CachedPostTest.java
@@ -5,14 +5,14 @@ import org.junit.Test;
 import static org.junit.Assert.*;
 
 import com.github.versus.offline.CachedPost;
+import com.github.versus.offline.SimpleTestPost;
 import com.github.versus.posts.Location;
 import com.github.versus.posts.Post;
 import com.github.versus.posts.Timestamp;
 
 import java.time.Month;
 import java.util.Calendar;
-import java.util.List;
-import java.util.Map;
+
 
 import static com.github.versus.offline.CachedPost.match;
 import static com.github.versus.offline.CachedPost.postIsInvalid;
@@ -26,151 +26,35 @@ public class CachedPostTest {
 
     @Test
     public void postIsInvalidWithNullTitle() {
-        Post post = new Post() {
-            @Override
-            public String getTitle() {
-                return null;
-            }
 
-            @Override
-            public Timestamp getDate() {
-                return new Timestamp(2022, Month.JANUARY, 1, 8, 1, Timestamp.Meridiem.PM);
-            }
+        Post post= SimpleTestPost.postWith(null,
+                new Timestamp(Calendar.getInstance().get(Calendar.YEAR)
+                        , Month.JANUARY, 1, 8, 1, Timestamp.Meridiem.PM),
+                new Location("Lausanne", 10, 10),10);
 
-            @Override
-            public Location getLocation() {
-                return new Location("Lausanne", 10, 10);
-            }
-
-            @Override
-            public List<Object> getPlayers() {
-                return null;
-            }
-
-            @Override
-            public int getPlayerLimit() {
-                return 10;
-            }
-
-            @Override
-            public Map<String, Object> getAllAttributes() {
-                return null;
-            }
-
-
-        };
         assertTrue(postIsInvalid(post));
     }
 
     @Test
     public void postIsInvalidWithNullLocation() {
-        Post post = new Post() {
-            @Override
-            public String getTitle() {
-                return "Testing Post";
-            }
 
-            @Override
-            public Timestamp getDate() {
-                return new Timestamp(2022, Month.JANUARY, 1, 8, 1, Timestamp.Meridiem.PM);
-            }
-
-            @Override
-            public Location getLocation() {
-                return null;
-            }
-
-            @Override
-            public List<Object> getPlayers() {
-                return null;
-            }
-
-            @Override
-            public int getPlayerLimit() {
-                return 10;
-            }
-
-            @Override
-            public Map<String, Object> getAllAttributes() {
-                return null;
-            }
-        };
+        Post post = SimpleTestPost.postWith("Invalid post",
+                new Timestamp(Calendar.getInstance().get(Calendar.YEAR), Month.JANUARY, 1, 8, 1, Timestamp.Meridiem.PM),
+                null,10);
         assertTrue(CachedPost.postIsInvalid(post));
     }
 
     @Test
     public void postIsInvalidWithNullDate() {
-        Post post = new Post() {
-            @Override
-            public String getTitle() {
-                return "Testing Post";
-            }
-
-            @Override
-            public Timestamp getDate() {
-                return null;
-            }
-
-            @Override
-            public Location getLocation() {
-                return new Location("Lausanne", 10, 10);
-            }
-
-            @Override
-            public List<Object> getPlayers() {
-                return null;
-            }
-
-            @Override
-            public int getPlayerLimit() {
-                return 10;
-            }
-
-            @Override
-            public Map<String, Object> getAllAttributes() {
-                return null;
-            }
-        };
+        Post post= SimpleTestPost.postWith("Invalid Post",null,
+                new Location("Lausanne", 10, 10),10);
         assertTrue(CachedPost.postIsInvalid(post));
     }
 
     @Test
     public void postIsInvalidWithValidPost() {
-        Post post = new Post() {
-            @Override
-            public String getTitle() {
-                return "Valid Post";
-            }
-
-            @Override
-            public Timestamp getDate() {
-                return new Timestamp(Calendar.getInstance().get(Calendar.YEAR), Month.JANUARY,1,8,1, Timestamp.Meridiem.PM);
-            }
-
-            @Override
-            public Location getLocation() {
-                return new Location("Lausanne", 10, 10);
-            }
-
-            @Override
-            public List<Object> getPlayers() {
-                return null;
-            }
-
-            @Override
-            public int getPlayerLimit() {
-                return 10;
-            }
-
-            @Override
-            public Map<String, Object> getAllAttributes() {
-                return null;
-            }
-
-
-        };
+        Post post = new SimpleTestPost();
         assertFalse(postIsInvalid(post));
-
     }
 
     @Test
@@ -180,37 +64,7 @@ public class CachedPostTest {
 
     @Test
     public void validCachedPostIsNotEmpty(){
-        Post post = new Post() {
-            @Override
-            public String getTitle() {
-                return "Valid Post";
-            }
-
-            @Override
-            public Timestamp getDate() {
-                return new Timestamp(Calendar.getInstance().get(Calendar.YEAR), Month.JANUARY, 1, 8, 1, Timestamp.Meridiem.PM);
-            }
-
-            @Override
-            public Location getLocation() {
-                return new Location("Lausanne", 10, 10);
-            }
-
-            @Override
-            public List<Object> getPlayers() {
-                return null;
-            }
-
-            @Override
-            public int getPlayerLimit() {
-                return 10;
-            }
-
-            @Override
-            public Map<String, Object> getAllAttributes() {
-                return null;
-            }
-        };
+        Post post = new SimpleTestPost();
         assertFalse(match(post).isEmpty);
     }
     @Test
@@ -224,40 +78,18 @@ public class CachedPostTest {
     }
     @Test
     public void computeIDWithValidPost(){
-        Post post = new Post() {
-            @Override
-            public String getTitle() {
-                return "Valid Post";
-            }
-
-            @Override
-            public Timestamp getDate() {
-                return new Timestamp(Calendar.getInstance().get(Calendar.YEAR), Month.JANUARY, 1, 8, 1, Timestamp.Meridiem.PM);
-            }
-
-            @Override
-            public Location getLocation() {
-                return new Location("Lausanne", 10, 10);
-            }
-
-            @Override
-            public List<Object> getPlayers() {
-                return null;
-            }
-
-            @Override
-            public int getPlayerLimit() {
-                return 10;
-            }
-
-            @Override
-            public Map<String, Object> getAllAttributes() {
-                return null;
-            }
-        };
-
+        Post post = new SimpleTestPost();
         assertTrue(computeID(post).equals(String.valueOf(post.hashCode())));
 
         }
+
+    @Test
+    public void revertValidPostsGivesBackTheSameFields(){
+        SimpleTestPost post = new SimpleTestPost();
+        CachedPost cached = CachedPost.match(post);
+        assertTrue(post.equals(cached.revert()));
+    }
+
+
 
 }
